좋아요 Jwt 추가

+++ controller/LikeController.js
@@ -1,34 +1,57 @@
 const conn = require("../mariadb");
 const { StatusCodes } = require("http-status-codes");
+const jwt = require("jsonwebtoken");
+const ensureAuthorization = require("../auth");
 
 const addLike = (req, res) => {
-  const { id } = req.params;
-  const { user_id } = req.body;
-  let sql = "INSERT INTO likes (user_id,liked_book_id) VALUES (?,?)";
-  let values = [user_id, id];
+  const book_id = req.params.id;
 
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
-    return res.status(StatusCodes.OK).json(results);
-  });
+  let authorization = ensureAuthorization(req, res);
 
-  res.json("좋아요 추가");
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. ",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "로그인 토큰이 유효하지 않습니다. ",
+    });
+  } else {
+    let sql = "INSERT INTO likes (user_id,liked_book_id) VALUES (?,?)";
+    let values = [authorization.id, book_id];
+
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
 
 const removeLike = (req, res) => {
-  const { id } = req.params;
-  const { user_id } = req.body;
-  let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?";
-  let values = [user_id, id];
+  const book_id = req.params.id;
+  let authorization = ensureAuthorization(req, res);
 
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
-    return res.status(StatusCodes.OK).json(results);
-  });
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. ",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "로그인 토큰이 유효하지 않습니다. ",
+    });
+  } else {
+    let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?";
+    let values = [authorization.id, book_id];
+
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
 
 module.exports = {
